-Added functionality to logout(Connected to firebase) -Added modifiers to screens -Added AppNavigation.kt
