+++ app/src/main/java/com/example/finalproject001/screen/mainmenu/AccountPage.kt
@@ -19,6 +19,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.example.finalproject001.Routes
+import com.google.firebase.Firebase
+import com.google.firebase.auth.auth
 
 @Composable
 fun AccountPage(navController: NavController, modifier: androidx.compose.ui.Modifier = Modifier){
@@ -41,9 +43,7 @@ fun AccountPage(navController: NavController, modifier: androidx.compose.ui.Modi
         Spacer(modifier = Modifier.height(5.dp))
 
         Button(onClick = {
-            navController.navigate(Routes.getstartedScreen){
-                popUpTo(Routes.getstartedScreen){inclusive = true}
-            }
+            Firebase.auth.signOut()
             Toast.makeText(context, "Successfully Logged out!", Toast.LENGTH_SHORT).show()
         }){
             Text(text = "Logout")
